# openapi-spec.yaml
openapi: "3.0.3"
info:
  version: 0.0.1
  title: Ponyhof API
  description: A small sample API to grasp the basics of the OpenAPI Specification
servers:
  - url: http://localhost:8080/api
paths:
  /games:
    get:
      summary: List all games
      operationId: GetGames
      responses:
        200:
          description: Array of Game objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
    post:
      summary: Add Game
      operationId: AddGame
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Game"
      responses:
        201:
          description: added the game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
    put:
      summary: Update a Game
      operationId: updateGame
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Game"
      responses:
        201:
          description: updated the game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
  /games/{id}:
    delete:
      summary: delete Game by id
      operationId: deleteGame
      parameters:
        - name: id
          in: path
          description: the id of the Publisher
          required: true
          schema:
            type: integer
            format: int32
      responses:
        200:
          description: successful
        400:
          description: No Game Found

  /publishers:
    get:
      summary: list all Publishers
      operationId: GetPublishers
      responses:
        200:
          description: Array of Publisher objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Publisher"
    post:
      summary: Add publisher
      operationId: AddPublisher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Publisher"
      responses:
        201:
          description: added the publisher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Publisher"
    put:
      summary: Update a publisher
      operationId: updatePublisher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Publisher"
      responses:
        201:
          description: updated the publisher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Publisher"

  /publishers/{id}:
    get:
      summary: Get Publisher by id
      operationId: GetPublisherById
      parameters:
        - name: id
          in: path
          description: the id of the Publisher
          required: true
          schema:
            type: integer
            format: int32
      responses:
        200:
          description: Get one Publisher object if found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Publisher"
        400:
          description: No publisher Found
    delete:
      summary: delete Publisher by id
      operationId: deletePublisher
      parameters:
        - name: id
          in: path
          description: the id of the Publisher
          required: true
          schema:
            type: integer
            format: int32
      responses:
        200:
          description: successful
        400:
          description: No publisher Found

  /ponies:
    get:
      summary: List all ponies
      operationId: getPonies
      parameters:
        - name: limit
          in: query
          description: How many items to return at one time
          required: false
          schema:
            type: integer
            format: int32
      responses:
        200:
          description: An paged array of pets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PonyDto"
    post:
      summary: introduce a new pony to the flock
      operationId: addPony
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PonyDto"
      responses:
        201:
          description: added the pony
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PonyDto"
components:
  schemas:
    Game:
      required:
        - Id
        - Name
        - PublisherIDs
        - Image
        - Rating
      properties:
        Id:
          type: integer
        Name:
          type: string
        Image:
          type: string
        ReleaseDate:
          type: string
          pattern: "^(3[01]|[12][0-9]|0[1-9])-(1[0-2]|0[1-9])-[0-9]{4}$"
          example: "2019-05-17"
        PublisherID:
          type: integer
        Rating:
          type: integer
        Downloads:
          type: integer
        Price:
          type: number
        Crossplay:
          type: boolean
        PlatformIDs:
          type: array
          items:
            type: integer
        Platforms:
          type: array
          items:
            $ref: "#/components/schemas/Platform"
        Category:
          $ref: "#/components/schemas/Category"
        Size:
          type: number
    Publisher:
      required:
        - Name
        - Id
      properties:
        Id:
          type: integer
        Name:
          type: string
        MarketCap:
          type: string
    Platform:
      required:
        - Id
        - Name
      properties:
        Id:
          type: integer
        Name:
          type: string
        Players:
          type: integer
    Category:
      type: string
      enum:
        - Action
        - Adventure
        - Fantasy
        - Shooter
    PonyDto:
      required:
        - name
        - birthday
      properties:
        id:
          type: integer
          example: 42
          minimum: 0
          readOnly: true
        name:
          type: string
          minLength: 1
          example: Thunder
        birthday:
          type: string
          format: date
          example: 2021-01-31